requirements.txt
FinMind==1.6.0
pandas==2.0.0
python-dotenv==1.0.0
pymongo==4.3.3
beautifulsoup4==4.11.2
requests==2.28.2
lxml>=4.9.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0
plotly>=5.18.0
